requirements.txt
Django==5.1.1
djangorestframework==3.15.2
Pillow==10.4.0
django-crispy-forms==2.3
crispy-bootstrap5
reportlab==4.2.5
python-decouple==3.8
django-extensions==3.2.3
django-widget-tweaks==1.5.1
python-docx==1.1.2
